+++ operator/internal/manifests/internal/config/build_test.go
@@ -90,6 +90,7 @@ limits_config:
   max_cache_freshness_per_query: 10m
   per_stream_rate_limit: 3MB
   per_stream_rate_limit_burst: 15MB
+  split_queries_by_interval: 30m
 memberlist:
   abort_if_cluster_join_fails: true
   bind_port: 7946
@@ -115,7 +116,6 @@ query_range:
       enable_fifocache: true
       fifocache:
         max_size_bytes: 500MB
-  split_queries_by_interval: 30m
   parallelise_shardable_queries: true
 schema_config:
   configs:
@@ -306,6 +306,7 @@ limits_config:
   max_cache_freshness_per_query: 10m
   per_stream_rate_limit: 3MB
   per_stream_rate_limit_burst: 15MB
+  split_queries_by_interval: 30m
 memberlist:
   abort_if_cluster_join_fails: true
   bind_port: 7946
@@ -331,7 +332,6 @@ query_range:
       enable_fifocache: true
       fifocache:
         max_size_bytes: 500MB
-  split_queries_by_interval: 30m
   parallelise_shardable_queries: true
 schema_config:
   configs: