+++ pkg/distributor/http_test.go
@@ -1,1 +1,59 @@
 package distributor
+
+import (
+	"context"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/grafana/dskit/flagext"
+	"github.com/grafana/dskit/services"
+	"github.com/stretchr/testify/require"
+
+	"github.com/grafana/loki/pkg/validation"
+)
+
+func TestDistributorRingHandler(t *testing.T) {
+	limits := &validation.Limits{}
+	flagext.DefaultValues(limits)
+
+	runServer := func() *httptest.Server {
+		d := prepare(t, limits, nil, nil)
+		defer services.StopAndAwaitTerminated(context.Background(), d) //nolint:errcheck
+
+		return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			d.ServeHTTP(w, r)
+		}))
+	}
+
+	t.Run("renders ring status for global rate limiting", func(t *testing.T) {
+		limits.IngestionRateStrategy = validation.GlobalIngestionRateStrategy
+		svr := runServer()
+		defer svr.Close()
+
+		resp, err := svr.Client().Get(svr.URL)
+		require.NoError(t, err)
+
+		defer resp.Body.Close()
+		body, err := ioutil.ReadAll(resp.Body)
+		require.NoError(t, err)
+		require.Contains(t, string(body), "<th>Instance ID</th>")
+		require.NotContains(t, string(body), "Not running with Global Rating Limit - ring not being used by the Distributor")
+	})
+
+	t.Run("doesn't return ring status for local rate limiting", func(t *testing.T) {
+		limits.IngestionRateStrategy = validation.LocalIngestionRateStrategy
+		svr := runServer()
+		defer svr.Close()
+
+		resp, err := svr.Client().Get(svr.URL)
+		require.NoError(t, err)
+
+		defer resp.Body.Close()
+		body, err := ioutil.ReadAll(resp.Body)
+		require.NoError(t, err)
+		require.Contains(t, string(body), "Not running with Global Rating Limit - ring not being used by the Distributor")
+		require.NotContains(t, string(body), "<th>Instance ID</th>")
+	})
+}