+++ pkg/storage/chunk/objectclient/client.go
@@ -22,19 +22,25 @@ var Base64Encoder = func(key string) string {
 	return base64.StdEncoding.EncodeToString([]byte(key))
 }
 
+const defaultMaxParallel = 150
+
 // Client is used to store chunks in object store backends
 type Client struct {
-	store      chunk.ObjectClient
-	keyEncoder KeyEncoder
-	schema     chunk.SchemaConfig
+	store               chunk.ObjectClient
+	keyEncoder          KeyEncoder
+	getChunkMaxParallel int
 }
 
 // NewClient wraps the provided ObjectClient with a chunk.Client implementation
-func NewClient(store chunk.ObjectClient, encoder KeyEncoder, schema chunk.SchemaConfig) *Client {
+func NewClient(store chunk.ObjectClient, encoder KeyEncoder) *Client {
+	return NewClientWithMaxParallel(store, encoder, defaultMaxParallel)
+}
+
+func NewClientWithMaxParallel(store chunk.ObjectClient, encoder KeyEncoder, maxParallel int) *Client {
 	return &Client{
-		store:      store,
-		keyEncoder: encoder,
-		schema:     schema,
+		store:               store,
+		keyEncoder:          encoder,
+		getChunkMaxParallel: maxParallel,
 	}
 }
 
@@ -87,7 +93,11 @@ func (o *Client) PutChunks(ctx context.Context, chunks []chunk.Chunk) error {
 
 // GetChunks retrieves the specified chunks from the configured backend
 func (o *Client) GetChunks(ctx context.Context, chunks []chunk.Chunk) ([]chunk.Chunk, error) {
-	return util.GetParallelChunks(ctx, chunks, o.getChunk)
+	getChunkMaxParallel := o.getChunkMaxParallel
+	if getChunkMaxParallel == 0 {
+		getChunkMaxParallel = defaultMaxParallel
+	}
+	return util.GetParallelChunks(ctx, getChunkMaxParallel, chunks, o.getChunk)
 }
 
 func (o *Client) getChunk(ctx context.Context, decodeContext *chunk.DecodeContext, c chunk.Chunk) (chunk.Chunk, error) {